PTable loading happens in static constructor
