+++ mzLib/UsefulProteomicsDatabases/Loaders.cs
@@ -190,16 +190,10 @@ namespace UsefulProteomicsDatabases
             return chargeMatch.Groups[1].Value;
         }
 
+        [Obsolete("This method is obsolete. It happens automatically on first call to periodic table in static constructor")]
         public static void LoadElements()
         {
-            // has the periodic table already been loaded?
-            if (PeriodicTable.GetElement(1) != null)
-            {
-                return;
-            }
 
-            // periodic table has not been loaded yet - load it
-            PeriodicTableLoader.Load();
         }
 
         public static IEnumerable<Modification> LoadUnimod(string unimodLocation)
